Add better project file support for virtualenv usage
Added code to find the project file and add $project_path var in
the settings file to prefix other paths. Also did this for
$working_dir. This is to avoid hard coded paths to virtualenvs
in your project settings. Getting the project file is straight
forward in ST3. In ST2, the code will search open folders to find
the first *.sublime-project file and load settings using JSON

* Add ability to use $project_path in settings paths
* Add ability to use $working_dir in settings paths